- moved tests from merged components to DefaultLayout tests - migrated tests - removed redundancy tests - fixed missing slot in DefaultLayout
